requirements.txt
openai
python-dotenv
mcp
proxies